refactor(App): remove unused components and update import paths
- Remove BadgeModal component and all related badge insertion functionality
- Remove CodeGeneratorModal component and its code generation logic
- Remove EditorLayout component and replace import to use components/layout
- Remove EmojiPicker component and its emoji selection interface
- Remove FolderModal component and folder structure generation UI
- Remove Footer component from the project UI
- Update App.tsx import statement for EditorLayout to new path components/layout
