+++ src/components/layout/index.ts
@@ -1,0 +1,5 @@
+export { EditorLayout } from './EditorLayout';
+export { Header } from './Header';
+export { Footer } from './Footer';
+export { Sidebar } from './Sidebar';
+export { StatusBar } from './StatusBar';
