+++ src/components/modals/BadgeModal.tsx
