+++ src/components/modals/index.ts
@@ -1,0 +1,7 @@
+export { default as BadgeModal } from './BadgeModal';
+export { default as TableModal } from './TableModal';
+export { default as FolderModal } from './FolderModal';
+export { default as MermaidModal } from './MermaidModal';
+export { default as CodeGeneratorModal } from './CodeGeneratorModal';
+export { default as ImageModal } from './ImageModal';
+export { default as TableOfContentsModal } from './TableOfContentsModal';
