+++ src/components/pickers/index.ts
@@ -1,0 +1,1 @@
+export { default as EmojiPicker } from './EmojiPicker';
