+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as MainContent } from './MainContent';
+export { default as MarkdownEditor } from './MarkdownEditor';
